fixed bug with type inference

+++ datasets/leetcode-all/LeetCode_327_numberOfBoomerangs.py
@@ -1,7 +1,7 @@
-from typing import List, Union
+from typing import Union, List
 
 
-def numberOfBoomerangs(points: List[Union[List[float], List[int]]]) -> int:
+def numberOfBoomerangs(points: List[Union[List[int], List[float]]]) -> int:
     """
     You are given n points in the plane that are all distinct, where points[i] = [xi, yi]. A boomerang is a tuple of points (i, j, k) such that the distance between i and j equals the distance between i and k (the order of the tuple matters).
     Return the number of boomerangs.
